feat(server): cache metabase responses on the exported singleton
Wrap the exported metabaseAPI singleton with createCachedMetabaseService
so getDashboard / findDashcard / getCardValue go through TTL-cached
fetches with promise coalescing. The cache is wired from config
(cacheTtlMs, cacheMaxEntries).

The previous shape — metabase-cache importing normalizeDashboard /
findDashcardRef from metabase-api, which in turn instantiates the cache
at module init — introduced a circular import that broke
metabase-cache.test.ts (CachedMetabaseService accessed before
initialization). Extract the pure normalization helpers into a new
metabase-normalize module so the dependency graph stays acyclic:
metabase-api -> metabase-cache -> metabase-normalize.

Add an afterEach to dashboard-api.test.ts that clears the cache so a
cached response from one test no longer leaks into the next.
